requirements.txt
matplotlib==3.10.8
numpy==2.4.4
ollama==0.6.1
PyYAML==6.0.3
tqdm==4.67.3
